Fix forward event times
Yeah, this one is complicated. To understand it, you kind of need to understand why the other approaches failed:

https://github.com/SciML/DiffEqSensitivity.jl/pull/435
https://github.com/SciML/DiffEqSensitivity.jl/pull/434

Essentially, you need the dual numbers in the callbacks otherwise there are cases where derivatives are not bounded by error tolerance at all. Period, no change whatsoever given the definition of the errors. But this means that your time stepping in the pure forward pass is different, and every chunk is slightly different to. That is kind of an issue for chunked forward mode when it's used in reverse mode, but we proceed. 

So the first idea was to track events to pull out the event times, but it turns out the thread-safety copy behavior makes this really hard. So the most sane solution seemed to be to discard the event times by checking non-uniqueness. In reality this is only a partial solution because someone could define events with save_positions=(true,false) or (false,true), and want to use that in a cost function. In practice, defining cost functions w.r.t. times which are not defined in advance is really... ill-defined? 

So this literally only shows up if the user somehow defined a cost function which uses whatever random t's the integrator spits out (so, not some fit against data), and if they've modified the callback so that it saves at only a single extra random point at each event. The only case I can think of even remotely sane for that is somehow a loss function on a Poincure diagram, but that's ill-defined because it's chaotic, so... I'm punting on that. What will happen in this case is someone will get an error saying that `du` or `dp` is not sized correctly to match the solution. If that error shows up in a user code we can handle it (possibly by just having them try UNPERTURBED_NORM), so at least this PR is safe and will throw an appropriate error in the bad case.

That said, this does show that chunked forward mode in reverse is a little ill-defined, and so that motivates bringing back a fully non-chunked mode. The things you learn from weird user examples...