Don't allow Zygote to return nothing by default
This is a small QOL change which makes it so ZygoteVJPs will throw an error if they have a `nothing` return, which is an implicit zero derivative due to not having parameters defined. This is the case where you have `f(u,p,t)` and then the user doesn't use `p`, for example `f(u,p,t) = NN(u)` instead of `re(p)(u)`. This sanity check will help users realize when they are accidentally using global variables instead of local variables and thus not differentiating w.r.t. `p`.